requirements.txt
addict==2.4.0
affine==2.4.0
aiofiles==22.1.0
aiosqlite==0.20.0
aniso8601==9.0.1
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.4
arcgis==2.3.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
backports.tarfile==1.2.0
basemap==1.4.1
basemap-data==1.3.2
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
bqplot==0.12.43
branca==0.7.1
cachelib==0.9.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==3.0.0
color-operations==0.1.3
colour==0.1.5
comm==0.2.1
contourpy==1.2.0
cryptography==42.0.7
cycler==0.12.1
cykhash==2.0.1
Cython==3.0.10
dask==2024.7.1
debugpy==1.8.1
decorator==5.1.1
deep-translator==1.11.4
defusedxml==0.7.1
dill==0.3.8
diot==0.2.3
duckdb==0.10.1
EasyNMT==2.0.2
executing==2.0.1
fastjsonschema==2.19.1
fasttext==0.9.2
filelock==3.13.1
fiona==1.9.6
Flask==3.0.2
Flask-Caching==2.1.0
Flask-Cors==4.0.0
flask-restx==1.3.0
folium==0.16.0
fonttools==4.49.0
fqdn==1.5.1
fsspec==2024.2.0
funda-scraper==1.1.1
gdown==5.1.0
geodatasets==2023.12.0
geojson==3.1.0
geomet==1.1.0
geopandas==0.14.3
google_streetview==1.2.9
googlemaps==4.10.0
googletrans==3.1.0a0
graphviz==0.20.3
groundingdino-py==0.4.0
gssapi==1.8.3
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.4.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.21.4
hyperframe==5.2.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
inflection==0.5.1
ipyevents==2.0.2
ipyfilechooser==0.6.0
ipykernel==6.29.3
ipyleaflet==0.18.2
ipython==8.22.1
ipython-genutils==0.2.0
ipytree==0.2.2
ipywidgets==7.8.2
isoduration==20.11.0
itsdangerous==2.1.2
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.18
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.3
jupyter-server==1.24.0
jupyter-ydoc==0.2.5
jupyter_core==5.7.1
jupyter_server_fileid==0.9.2
jupyter_server_terminals==0.5.2
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.7
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.3
jupyterlab_widgets==1.1.8
keyring==25.2.1
kiwisolver==1.4.5
krb5==0.6.0
kwconfig==1.1.7
leafmap==0.31.5
llvmlite==0.42.0
localtileserver==0.10.1
locket==1.0.0
lxml==5.2.2
mapclassify==2.6.1
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mistune==3.0.2
more-itertools==10.3.0
morecantile==5.3.0
mpmath==1.3.0
nbclassic==1.1.0
nbclient==0.9.0
nbconvert==7.16.1
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
notebook==6.5.7
notebook_shim==0.2.4
numba==0.59.1
numexpr==2.9.0
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.9.0.80
osmnx==1.9.1
overrides==7.7.0
packaging==23.2
pandarallel==1.6.5
pandas==2.1.4
pandocfilters==1.5.1
parso==0.8.3
partd==1.4.2
patsy==0.5.6
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
plotly==5.19.0
pooch==1.8.1
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==5.26.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
puremagic==1.26
py-cpuinfo==9.0.0
pyarrow==17.0.0
pybind11==2.12.0
pycocotools==2.0.7
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
pylerc==4.0
pyparsing==3.1.1
pyproj==3.6.1
pyrobuf==0.9.3
pyrosm==0.6.2
pyshp==2.3.1
PySocks==1.7.1
pyspnego==0.10.2
pystac==1.9.0
pystac-client==0.7.6
python-box==7.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-rapidjson==1.17
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
rasterio==1.3.9
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-gssapi==1.3.0
requests-kerberos==0.15.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rio-cogeo==5.2.0
rio-tiler==6.4.3
rpds-py==0.18.0
ruff==0.4.4
sacremoses==0.1.1
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
scooby==0.9.2
seaborn==0.13.2
segment-anything-hq==0.3
segment-anything-py==1.0
segment-geospatial==0.10.2
Send2Trash==1.8.2
sentencepiece==0.2.0
server-thread==0.2.0
shap==0.45.0
shapely==2.0.3
six==1.16.0
slicer==0.0.7
smbprotocol==1.13.0
sniffio==1.3.1
snuggs==1.4.7
soupsieve==2.5
stack-data==0.6.3
statsmodels==0.14.1
supervision==0.6.0
sympy==1.12
tenacity==8.2.3
terminado==0.18.0
thop==0.1.1.post2209072238
threadpoolctl==3.3.0
timm==0.9.16
tinycss2==1.2.1
tokenizers==0.15.2
tomli==2.0.1
toolz==0.12.1
torch==2.2.1
torchaudio==2.2.1
torchvision==0.17.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.1
traittypes==0.2.1
transformers==4.38.2
truststore==0.9.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.10.0
tzdata==2024.1
ujson==5.10.0
ultralytics==8.1.24
uri-template==1.3.0
urllib3==2.2.2
uvicorn==0.29.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
whitebox==2.3.1
whiteboxgui==2.3.0
widgetsnbextension==3.6.7
xarray==2024.2.0
xgboost==2.0.3
xyzservices==2023.10.1
y-py==0.6.2
yapf==0.40.2
ypy-websocket==0.8.4
zipp==3.18.1
